scripts/requirements.txt
# Requirements for note normalization scripts
psycopg2-binary>=2.9.0
pyyaml>=6.0
requests>=2.28.0  # For AI description generation

# FastAPI service
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
